services/image-processing-service/requirements.txt
# 图像处理微服务核心依赖 - 无状态架构
# 专注于图像处理算法，数据存储通过storage-service完成

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP客户端（与storage-service通信）
httpx==0.25.2
aiofiles==23.2.0

# 核心图像处理库
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
imageio==2.31.5

# 深度学习和计算机视觉
torch==2.1.1
torchvision==0.16.1
numpy==1.24.4

# 图像增强和处理
albumentations==1.3.1
imgaug==0.4.0

# 科学计算
scipy==1.11.4
scikit-learn==1.3.2

# 文档图像处理专用
# deskew==1.3.3  # 倾斜校正专用库

# 图像质量评估
# lpips==0.1.4  # 感知相似性
# pytorch-msssim==1.0.0  # 结构相似性

# 数据验证和配置
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志和监控
loguru==0.7.2